Renamed 'configuration' to mic_setup for clarity's sake
