+++ profile/src/microphone.rs
@@ -1,5 +1,5 @@
 pub mod equalizer;
 pub mod compressor;
 pub mod gate;
-pub mod configuration;
+pub mod mic_setup;
 pub mod ui_setup;
